added search by barcode -- test pass

+++ Repositories/Items/IItemsRepository.cs
@@ -9,6 +9,6 @@ namespace PdaHub.Repositories.Items
         Task<ItemSectionEnitiyModel> GetItemSectionAsync(string barcode);
         Task<PosItemEnitityModel> GetPosItemAsync(string barcode);
         Task<ItemSpecialEnitiyModel> ItemSpecialAsync(string barcode);
-        Task<List<PosItemEnitityModel>> GetItemsInPromo(int DiscountNo);
+        Task<List<PosItemEnitityModel>> GetItemsInPromo(long DiscountNo);
     }
 }
